+++ website/service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ec-lernstudio-v37";
+const CACHE_NAME = "ec-lernstudio-v38";
 const APP_SHELL = [
   "./",
   "./index.html",